requirements.txt
anyio==4.6.2.post1
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ollama==0.3.3
sniffio==1.3.1
typing_extensions==4.12.2
beautifulsoup4==4.12.3
PyQt5==5.15.11
python-dotenv==1.0.1

requests==2.32.3
vvspy==2.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.0 
SpeechRecognition==3.11.0
pyttsx3==2.98
transitions==0.9.2

newsapi-python==0.2.7
regex==2022.9.13
# Add Windows-only packages in this format
win10toast==0.9 ; platform_system == "Windows"
pypiwin32 ; platform_system == "Windows"